+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 ## [0.1.0] - 19/11/2025
 
+### Added
+ - US for modify supply batch
+ - US for display of supply batches based on date and idSupply
+ - US for modify User with MVVM and clean architecture principles.
+
 ### Fixed
 - Fixed search bar in workshop list.
+- Fixed the product error message
+- Error now properly appearing when registering new beneficiary
+- Fixed error views in product batches list, detail, register and modify
+- Fixed product fetch in register Product Batch view
 
 ## [0.1.0] - 18/11/2025